@@ -4,15 +4,15 @@ version: 1.2.1
 homepage: https://github.com/daniloapr/cubos_widgets
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
   flutter: ">=1.17.0"
 
 dependencies:
   flutter:
     sdk: flutter
   
-  url_launcher: ^6.0.2
-  flutter_html: ^1.3.0
+  url_launcher: ^6.0.3
+  flutter_html: ^2.0.0
 
 dev_dependencies:
   flutter_test:
